auto-record_case-classification/requirements.txt
# Function dependencies, for example:
# package>=version

line-bot-sdk
google-cloud-storage
google-cloud-firestore